% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_calc.R
\name{alpha.calc}
\alias{alpha.calc}
\title{Calculates alpha via a simple function}
\usage{
alpha.calc(chart, mapping, f.alpha)
}
\arguments{
\item{chart}{The **data.frame** generated by
diff.broadcast}

\item{mapping}{The **data.frame** generated by
create.move.mapping}

\item{f.alpha}{The **function** to calculate alpha
with (diffs, moves.values) respectively}
}
\description{
Doesn't really need to exist as a separate function
but it acts as a wrapper
}
