% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_parser.R
\name{chart.parse}
\alias{chart.parse}
\title{Parses the chart into a data.frame}
\usage{
chart.parse(chart.path = NA, chart.lines = NA)
}
\arguments{
\item{chart.path}{Path of the chart to be parsed}

\item{chart.lines}{Lines of the chart to be parsed, if the chart
is not in a file format}

\item{chart}{The chart to be parsed, in a vector of characters.
This can be provided via readLines function.}
}
\value{
A data.frame consisting of the note's data only.
Columns: keys, types, offsets
}
\description{
Only .osu formats are supported for now.
}
