% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartFngMapping.R
\name{chartFngMapping}
\alias{chartFngMapping}
\title{Creates the mapping to be used alongside with
the broadcasted chart.}
\usage{
chartFngMapping(keyset.select = NA, keyset = NA, include.details = T)
}
\arguments{
\item{keyset.select}{A character vector based on
type of keyset.

Either keyset or keyset.select must be defined

Valid Keyset.select values are

'4', '5L', '5R', '6', '7L', '7R',
'8SPR', '8SPL','8SYM', '9L', '9R'

Only 8K has special keysets

8SYM: 8 Key Symmetry

8SPL: 8 Key Special (Left)

8SPR: 8 Key Special (Right)}

\item{keyset}{A data.frame based on type of keyset.

Either keyset or keyset.select must be defined

Fingers are denoted as
P|inky
R|ing
M|iddle
I|ndex
T|humb

Column keys and fingers is required.

An example of keyset would be

7R = data.frame(keys = 1:7,
                fingers = c(2,3,4,6,7,8,9))}

\item{include.details}{A Logical to indicate if
details should be calculated

Directions

Directions is a character vector indicating what
type of motion it is.

"across" means moves from hand to hand

"jack" is self-explanatory

"in" is where the motion is leading towards the
thumb

"out" is where the motion is leading to the pinkies

Distance is the number of columns between the pair of
notes

rfls is the ave distance from the center, (5.5).
rfls stands for reflections, as it's the method used to
calculate.}
}
\description{
Using the broadcasted format, this function maps
the values provided by `mapping`

This can also be skipped if you're doing a custom
mapping
}
