% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMoveMapping.R
\name{createMoveMapping}
\alias{createMoveMapping}
\title{Creates the mapping to be used alongside with
the broadcasted chart.}
\usage{
createMoveMapping(keyset.select = NA, keyset = NA, mapping = NA,
  mapping.opp = NA)
}
\arguments{
\item{keyset.select}{A **character vector** based on
type of keyset.

Valid Keyset.select values are '4','5', ..., '8SP','8SY','9'

Only 8K has special keysets

8SYM: 8 Key Symmetry

8SPL: 8 Key Special (Left)

8SPR: 8 Key Special (Right)}

\item{keyset}{A **data.frame** based on type of keyset.

Fingers are denoted as
P|inky
R|ing
M|iddle
I|ndex
T|humb

Column keys and fingers is required.

An example of keyset would be

'7' = data.frame(keys = 1:7,
                 fingers = c('R','M','I','T','I','M','R'))}

\item{mapping}{A custom **5 by 5 matrix** to be used
to map onto the finger. If NA, a default would be used

An example of a mapping. Values can change, but not
the format.

pr means pinky to ring, rm means ring to middle, ...

# [P] [R] [M] [I] [T]

c(7.0,5.0,3.0,2.7,2.7), # [P] pp,pr,pm,pi,pt

c(6.0,6.0,2.5,2.0,2.0), # [R] rp,rr,rm,ri,rt

c(3.5,3.0,5.0,1.0,1.3), # [M] mp,mr,mm,mi,mt

c(2.7,1.2,1.2,5.0,2.4), # [I] ip,ir,im,ii,it

c(2.7,1.7,1.5,1.9,5.0)  # [T] tp,tr,tm,ti,tt}

\item{mapping.opp}{A custom **5 by 5 matrix** to be used
to map onto actions that involve alternating hands.
If NA, a default would be used

Same format as `mapping`. However, pr would mean
FROM Hand 1 Pinky to Hand 2 Ring. Rest is the same idea.}
}
\description{
Using the broadcasted format, this function maps
the values provided by `mapping`

This can also be skipped if you're doing a custom
mapping
}
