% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffBroadcast.R
\name{diffBroadcast}
\alias{diffBroadcast}
\title{For each row, gets the time before the next key}
\usage{
diffBroadcast(chart, ignore.types = c("lnotel"))
}
\arguments{
\item{chart}{The data.frame generated by f.chart.parse}

\item{ignore.types}{A character vector determining what
types of notes should be ignored}
}
\description{
This will not overwrite the current columns,
it will generate a separate one.
}
