% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffCalc.R
\name{diffCalc}
\alias{diffCalc}
\title{Calculates the difficulty}
\usage{
diffCalc(osu.stress, osu.bcst, f.merge.key, f.merge.bin, bin.size = 5000)
}
\arguments{
\item{osu.stress}{Get this data.frame from stressSim}

\item{osu.bcst}{Get this data.frame from diffBroadcast}

\item{f.merge.key}{Specify the function used to merge
key values together}

\item{f.merge.bin}{Specify the function used to merge
bins values together}

\item{bin.size}{Size of the bin}
}
\description{
Calculates the difficulty
}
