% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cppModelLongNote}
\alias{.cppModelLongNote}
\title{Helps model.longNote in speeding up operations}
\usage{
.cppModelLongNote(chart, longNoteStartName, longNoteEndName, newName,
  labelStart, labelEnd)
}
\arguments{
\item{chart}{A casted chart provided via model.longNote}

\item{longNoteStartName}{The string indicating the start of a long note}

\item{longNoteEndName}{The string indicating the start of a long note}

\item{newName}{The string indicating the replaced name of long note bodies}

\item{labelStart}{A boolean indicating if the start of the long notes must
be labelled}

\item{labelEnd}{A boolean indicating if the end of the long notes must be
labelled}
}
\description{
Helps model.longNote in speeding up operations
}
