% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cppModelManipulation}
\alias{.cppModelManipulation}
\title{Helps model.manipulation in speeding up operations}
\usage{
.cppModelManipulation(unq_offsets, obj_offsets, window = 200,
  is_sorted = FALSE)
}
\description{
Calculates how many notes in front of
window
}
