% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_defaultModelDensityMapping.R
\name{.dflt.model.density.mapping}
\alias{.dflt.model.density.mapping}
\title{Default Model Mapping for model.density}
\usage{
.dflt.model.density.mapping()
}
\description{
Default Model Mapping for model.density
}
