% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfltMoveKeysets.R
\name{.dflt.move.keysets}
\alias{.dflt.move.keysets}
\title{Default Move Keyset for chartFngMapping}
\usage{
.dflt.move.keysets()
}
\description{
Note that those with multiple playstyle choices have a default mapping,
usually defaulted to Right Hand Playstyle.
}
