% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelJack.R
\name{model.jack}
\alias{model.jack}
\title{Static Model generator for Jack}
\usage{
model.jack(chart.ext)
}
\arguments{
\item{chart.ext}{The chart generated from chartBroadcast}
}
\value{
Returns a data.frame compatible with model.sim
}
\description{
This uses chartBroadcast and just inverses the difference between
equal keys.froms and keys.tos
}
