% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelJoin.R
\name{model.join}
\alias{model.join}
\title{Joins all models into a joined tabular format}
\usage{
model.join(m.mtn, m.dns, m.mnp, m.lng)
}
\arguments{
\item{m.mtn}{model generated by model.motion}

\item{m.dns}{model generated by model.density}

\item{m.mnp}{model generated by model.manipulation}

\item{m.lng}{model generated by model.longNote}
}
\description{
Joins all models into a joined tabular format
}
