% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLongNote.R
\name{model.longNote}
\alias{model.longNote}
\title{Static Model generator for Long Notes}
\usage{
model.longNote(chart, chart.keyset.select = NA, chart.keyset = NA,
  directions.mapping = NA, scale = 1)
}
\arguments{
\item{chart}{The chart generated from chartParse}

\item{chart.keyset.select}{Read ?createMoveMapping for more details}

\item{chart.keyset}{Read ?createMoveMapping for more details}

\item{directions.mapping}{A data.frame to be merged with the output
directions to generate weights.

It must hold the columns directions and weights

If NA, .dflt.model.motion.mapping will be used}
}
\description{
Note that we define difficulty by the Long Note to Note Movement, so the
froms are Long Notes, tos are Notes.

In other words, if you're holding X, we define difficulty for a Y as
X -> Y akin to model.motion.
}
