% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelManipulation.R
\name{model.manipulation}
\alias{model.manipulation}
\title{A model that detects manipulatability in charts}
\usage{
model.manipulation(chart, window = 1000, bias.power = 2,
  ignore.types = c("lnotel"))
}
\arguments{
\item{chart}{chart generated by chartParse}

\item{window}{The window to check for biases.}

\item{bias.power}{The power value of the bias. A higher value will suppress
lower values more}

\item{ignore.types}{Types to be ignored during model creation}
}
\description{
Manipulation is where certain notes can be played in an
incorrect order while still maintaining a good accuracy and judgement.

This model aims to look at notes forward of each other and calculates a value
called a bias.

Biases are how the chart decides to favor a specific column over the other by
having higher or lower density.

The higher the bias, the less manipulatable it'll be
}
