% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMotion.R
\name{model.motion}
\alias{model.motion}
\title{Static Model generator for Motion}
\usage{
model.motion(chart.bcst, move.mapping = NA, keyset.select = NA,
  suppress.low.diff = T, suppress.threshold = 50, suppress.power = 2)
}
\arguments{
\item{chart.bcst}{The chart generated from diffBroadcast}

\item{move.mapping}{The data.frame generated from
createMoveMapping, if NA, keyset.select must be defined
to generate a default move.mapping}

\item{keyset.select}{If move.mapping is NA, this must be
specified to generate a default move.mapping. For valid
values refer to ?createMoveMapping}
}
\description{
This uses diffBroadcast and createMoveMapping
to
}
