% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSim.R
\name{model.sim}
\alias{model.sim}
\title{Simulates the Models through types}
\usage{
model.sim(m.jck, m.mtn, m.dns)
}
\arguments{
\item{m.jck}{model generated by model.jackInv}

\item{m.mtn}{model generated by model.motion}

\item{m.dns}{model generated by model.density}
}
\description{
Simulates the Models through types
}
