% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replay_parser.R
\name{replayParse}
\alias{replayParse}
\title{Parses the replay generated through the Python API downloader.}
\usage{
replayParse(chart, replay.path, ignore.threshold = 100)
}
\arguments{
\item{chart}{A **data.frame** that needs to have types, offsets,
and keys}

\item{replay.path}{**Path** of the replay generated.}

\item{ignore.threshold}{**Numeric** to indicate what deviations should
be ignored.}
}
\value{
Returns a joined **data.frame** with replay.offsets
on the following column.
}
\description{
Using proximity matching, data of the replay is
joined with the chart, so we can tell what notes are the replay
related to.
}
