% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_sim.R
\name{stress.sim}
\alias{stress.sim}
\title{Simulates the Stress through types}
\usage{
stress.sim(chart, stress.init = 0, decay_alpha = 1.5,
  decay_beta = 1000)
}
\arguments{
\item{chart}{Chart generated by `f.stress.mapper`}

\item{stress.init}{The **numeric** of 'stress' to
initialize the simulation with}
}
\description{

}
