% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stressSim.R
\name{stressSim}
\alias{stressSim}
\title{Simulates the Stress through types}
\usage{
stressSim(chart, stress.init = 0, decay.ms = 0.001)
}
\arguments{
\item{chart}{Chart generated by `f.stress.mapper`}

\item{stress.init}{The **numeric** of 'stress' to
initialize the simulation with}

\item{decay.ms}{The amount of stress decay per ms}
}
\description{
Simulates the Stress through types
}
