% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_transfer.R
\name{stressTransfer}
\alias{stressTransfer}
\title{Transfers stress to other keys}
\usage{
stressTransfer(chart, transfer.group, func)
}
\arguments{
\item{chart}{The **data.frame** generated by f.chart.sim}

\item{func}{A **function** to use on the group. Refer to
examples for specifications.}

\item{tf.group}{A **list** of key groups to have internal
stress transfer. There's a specific format to use.}
}
\description{
Depending on the function provided,
stress transfer may vary. However, in most cases
these stress transfers are assumptions on playstyle

This will not overwrite the current stress columns,
it will generate a separate one.
}
\examples{
The function also has a specific format
`function(stress.main, stress.others){ return(stress.main) }`
`stress.main` is the current key's stress value
`stress.others` is the other key's stress values in a vector.

The function must return the stress value.

}
