% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDifficulty.R
\name{calculateDifficulty}
\alias{calculateDifficulty}
\title{Calculates difficulty using stressSim}
\usage{
calculateDifficulty(chart.path = NA, chart.lines = NA,
  keyset.select = "4", keyset = NA, mtn.suppress = T,
  mtn.suppress.threshold = 50, mtn.suppress.scale = 2,
  mtn.across.weight = 0.7, mtn.in.weight = 1, mtn.out.weight = 1.3,
  mtn.jack.weight = 3, dns.window = 1000, dns.mini.ln.parse = T,
  dns.mini.ln.threshold = 150, dns.mini.ln.tail.drop = T,
  dns.note.weight = 1, dns.lnoteh.weight = 1, dns.lnotel.weight = 1,
  dns.m.lnoteh.weight = 1, dns.m.lnotel.weight = 1, decay.ms = 0.5,
  stress.init = 0)
}
\arguments{
\item{chart.path}{Path to chart}

\item{chart.lines}{ReadLines on the path}

\item{keyset.select}{A character vector indicating
the default keyset to map to the keys
Refer to ?chartFngMapping for more details}

\item{keyset}{A custom data.frame indicating the
default keyset to map to the keys
Refer to ?chartFngMapping for more details}

\item{mtn.suppress}{A logical indicating if small
motions should be suppressed
Refer to ?model.motion}

\item{mtn.suppress.threshold}{A numeric indicating
the threshold of what determines a "small motion"
Refer to ?model.motion}

\item{mtn.suppress.scale}{A numeric indicating the
power of suppression on small motions
Refer to ?model.motion}

\item{mtn.across.weight}{A numeric indicating the
weight of the 'across' motion weight
Refer to ?model.motion}

\item{mtn.in.weight}{A numeric indicating the
weight of the 'in' motion weight
Refer to ?model.motion}

\item{mtn.out.weight}{A numeric indicating the
weight of the 'out' motion weight
Refer to ?model.motion}

\item{mtn.jack.weight}{A numeric indicating the
weight of the 'jack' motion weight
Refer to ?model.motion}

\item{dns.window}{A numeric indicating the window to
check for density
Refer to ?model.density}

\item{dns.mini.ln.parse}{A logical indicating if mini
Long Note should be parsed
Refer to ?model.density}

\item{dns.mini.ln.threshold}{A numeric indicating the
threshold of what determines a "mini LN"
Refer to ?model.density}

\item{dns.mini.ln.tail.drop}{A logical indicating if
mini LNs should be dropped.

It is recommended to drop mini LNs for density
calculation

Refer to ?model.density}

\item{dns.note.weight}{A numeric indicating the
weight of the 'note' density weight
Refer to ?model.density}

\item{dns.lnoteh.weight}{A numeric indicating the
weight of the 'lnoteh' density weight
Refer to ?model.density}

\item{dns.lnotel.weight}{A numeric indicating the
weight of the 'lnotel' density weight
Refer to ?model.density}

\item{dns.m.lnoteh.weight}{A numeric indicating the
weight of the 'm.lnoteh' density weight
Refer to ?model.density}

\item{dns.m.lnotel.weight}{A numeric indicating the
weight of the 'm.lnotel' density weight
Refer to ?model.density}

\item{decay.ms}{A numeric indicating the decay of
stress per ms
Refer to ?model.sim}

\item{stress.init}{A numeric indicating the starting
stress
Refer to ?model.sim}
}
\description{
Calculates difficulty using stressSim
}
