% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDifficulty.R
\name{calculateDifficulty}
\alias{calculateDifficulty}
\title{Processes the file into difficulty models.}
\usage{
calculateDifficulty(chart.path = NA, chart.lines = NA,
  chart.rate = 1, chart.keyset.select = "4", chart.keyset = NA,
  mtn.suppress = T, mtn.suppress.threshold = 50,
  mtn.suppress.scale = 2, mtn.ignore.jacks = T,
  mtn.across.weight = 0.7, mtn.in.weight = 1, mtn.out.weight = 1.3,
  mtn.jack.weight = 3, dns.window = 1000, dns.mini.ln.parse = T,
  dns.mini.ln.threshold = 150, dns.mini.ln.tail.drop = T,
  dns.note.weight = 1, dns.lnoteh.weight = 1, dns.lnotel.weight = 1,
  dns.m.lnote.weight = 1, dns.m.lnotel.weight = 1, mnp.window = 1000,
  mnp.bias.power = 2, sim.decay.ms = 0.5, sim.stress.init = 0,
  sim.mtn.pow = 1, sim.dns.pow = 1, sim.bin.size = 5000)
}
\arguments{
\item{chart.path}{Path to chart}

\item{chart.lines}{ReadLines on the path}

\item{chart.rate}{The rate to judge the chart on}

\item{chart.keyset.select}{A character vector indicating the default
chart.keyset to map to the keys.

Refer to ?chartFngMapping for more details}

\item{chart.keyset}{A custom data.frame indicating the default chart.keyset
to map to the keys.

Refer to ?chartFngMapping for more details}

\item{mtn.suppress}{A logical indicating if small motions should be
suppressed

Refer to ?model.motion}

\item{mtn.suppress.threshold}{A numeric indicating the threshold of what
determines a "small motion".

Refer to ?model.motion}

\item{mtn.suppress.scale}{A numeric indicating the power of suppression on
small motions

Refer to ?model.motion}

\item{mtn.ignore.jacks}{A logical indicating if jacks should be ignored by
mtn and handled separately by JackInv

Refer to ?model.motion}

\item{mtn.across.weight}{A numeric indicating the weight of the 'across'
motion weight

Refer to ?model.motion}

\item{mtn.in.weight}{A numeric indicating the weight of the 'in' motion
weight

Refer to ?model.motion}

\item{mtn.out.weight}{A numeric indicating the weight of the 'out' motion
weight

Refer to ?model.motion}

\item{mtn.jack.weight}{A numeric indicating the weight of the 'jack' motion
weight

Refer to ?model.motion}

\item{dns.window}{A numeric indicating the window to check for density

Refer to ?model.density}

\item{dns.mini.ln.parse}{A logical indicating if mini Long Note should be
parsed.

Refer to ?model.density}

\item{dns.mini.ln.threshold}{A numeric indicating the threshold of what
determines a "mini LN".

Refer to ?model.density}

\item{dns.mini.ln.tail.drop}{A logical indicating if mini LNs should be
dropped.

It is recommended to drop mini LNs for density calculation.

Refer to ?model.density}

\item{dns.note.weight}{A numeric indicating the weight of the 'note' density
weight.

Refer to ?model.density}

\item{dns.lnoteh.weight}{A numeric indicating the weight of the 'lnoteh'
density weight.

Refer to ?model.density}

\item{dns.lnotel.weight}{A numeric indicating the weight of the 'lnotel'
density weight.

Refer to ?model.density}

\item{dns.m.lnote.weight}{A numeric indicating the weight of the 'm.lnote'
density weight.

Refer to ?model.density}

\item{dns.m.lnotel.weight}{A numeric indicating the weight of the 'm.lnotel'
density weight.

Refer to ?model.density}

\item{mnp.window}{A numeric indicating the window to check for bias.

Refer to ?model.manipulation}

\item{mnp.bias.power}{A numeric indicating the intensity to suppress lower
biases.

Refer to ?model.manipulation}

\item{sim.decay.ms}{A numeric indicating the decay of stress per ms.

Refer to ?model.sim}

\item{sim.stress.init}{A numeric indicating the starting stress.

Refer to ?model.sim}

\item{sim.mtn.pow}{A numeric indicating the power weight of the Motion model.

Refer to ?model.sim}

\item{sim.dns.pow}{A numeric indicating the power weight of the Density
model.

Refer to ?model.sim}

\item{sim.bin.size}{A numeric indicating the size of model.sim bin size.

Refer to ?model.sim}
}
\value{
Returns a list of calculated models

sim: Simulated, model: merged model, ...
}
\description{
calculateDifficulty(chart.path = "../7Kdifficulty.osu",
chart.keyset.select = '7R')

calculateDifficulty(chart.path = "../4Kdifficulty.osu",
chart.keyset.select = '4')
}
