% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartExtract.R
\name{chartExtract}
\alias{chartExtract}
\title{Extracts chart details to use with Models}
\usage{
chartExtract(chart, keyset.select = NA, keyset = NA,
  ignore.types = c("lnotel"))
}
\arguments{
\item{chart}{Chart generated by chartParse}

\item{keyset.select}{Read ?createMoveMapping
for more details}

\item{keyset}{Read ?createMoveMapping
for more details}

\item{ignore.types}{Types of notes to ignore

?diffBroadcast for more details}
}
\value{
Returns a data.frame compatible with
models
}
\description{
This acts as a wrapper for more
detailed functions. Outputs chart.ext.
}
