% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartExtract.R
\name{chartExtract}
\alias{chartExtract}
\title{Extracts chart details to use with Models}
\usage{
chartExtract(chart, keyset.select = NA, keyset = NA)
}
\arguments{
\item{chart}{Chart generated by chartParse}

\item{keyset.select}{Read ?createMoveMapping
for more details}

\item{keyset}{Read ?createMoveMapping
for more details}
}
\description{
This acts as a wrapper for more
detailed functions. Outputs chart.ext.
}
