% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.cppSimulateKey}
\alias{.cppSimulateKey}
\title{Simulates stress using a offset and value vector}
\usage{
.cppSimulateKey(offsets, values, decay_ms = 0.01, stress_init = 0)
}
\arguments{
\item{offsets}{Offsets of the values used to indicate spikes}

\item{values}{Spike values used to increase stress}

\item{decay_ms}{Stress decay per ms}

\item{stress_init}{Initial Stress}
}
\description{
Simulates stress using a offset and value vector
}
