% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDensity.R
\name{model.density}
\alias{model.density}
\title{Static Model generator for Density}
\usage{
model.density(chart, window = 1000, mini.longnote.parse = T,
  mini.longnote.threshold = 150)
}
\arguments{
\item{chart}{The chart generated from chartParse}

\item{window}{The window to check for objects}
}
\description{
Density is not meant to be a representation
of physical stress, instead it is for mental stress.

In other words, the density of objects within a certain
time frame to be processed.
}
