% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDensity.R
\name{model.density}
\alias{model.density}
\title{Static Model generator for Density}
\usage{
model.density(chart, window = 1000, mini.ln.len.min = 100,
  mini.ln.len.max = 400, mini.ln.weight.min = 0.65,
  mini.ln.weight.max = 1, weight.note = 1)
}
\arguments{
\item{chart}{The chart generated from chartParse}

\item{window}{The window to check for objects}

\item{mini.ln.len.min}{Defines the minimum length of a Mini Long Note}

\item{mini.ln.len.max}{Defines the maximum length of a Mini Long Note}

\item{mini.ln.weight.max}{Defines the minimum weight of a Mini Long Note

Note that the weight counts separately for the head and tail. So 0.5 will
treat a Long Note as 1 weight}

\item{weight.note}{Defines the weight of a note}

\item{mini.ln.weight.max}{Defines the maximum weight of a Mini Long Note

Note that the weight counts separately for the head and tail. So 0.5 will
treat a Long Note as 1 weight}
}
\value{
Returns a data.frame compatible with model.sim
}
\description{
Density is not meant to be a representation of physical stress,
instead it is for mental stress.

In other words, the density of objects within a certain time frame to be
processed.
}
