% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDensity.R
\name{model.density}
\alias{model.density}
\title{Static Model generator for Density}
\usage{
model.density(chart.ext, window = 1000, mini.ln.parse = T,
  mini.ln.threshold = 150, mini.ln.tail.drop = T, types.mapping = NA)
}
\arguments{
\item{chart.ext}{The chart generated from chartExtract}

\item{window}{The window to check for objects}

\item{mini.ln.parse}{A logical indicating if miniLNs
should be parsed separately}

\item{mini.ln.threshold}{A numeric indicating the
threshold of LNs to be considered mini}

\item{mini.ln.tail.drop}{Logical indicating if the
miniLN Tail should be dropped.}

\item{types.mapping}{A data.frame to be merged with the
output types to generate weights.

It must only hold the columns types and weights

If NA, .dflt.model.density.mapping will be used}
}
\description{
Density is not meant to be a representation
of physical stress, instead it is for mental stress.

In other words, the density of objects within a certain
time frame to be processed.
}
