% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelJackInv.R
\name{model.jackInv}
\alias{model.jackInv}
\title{Static Model generator for Jack Inverse}
\usage{
model.jackInv(chart.bcst)
}
\arguments{
\item{chart.bcst}{The chart generated from diffBroadcast}
}
\description{
This uses diffBroadcast and just inverses
the difference between equal keys.froms and keys.tos
}
