% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSim.R
\name{model.sim}
\alias{model.sim}
\title{Simulates the Models through types}
\usage{
model.sim(m.mtn, m.dns, m.mnp, mtn.pow, dns.pow, decay.ms = 0,
  decay.perc.s = 0.1, stress.init = 0, bin.size = 5000,
  sim.disable = F)
}
\arguments{
\item{m.mtn}{model generated by model.motion}

\item{m.dns}{model generated by model.density}

\item{m.mnp}{model generated by model.manipulation}

\item{mtn.pow}{power factor for mtn}

\item{dns.pow}{power factor for dns}

\item{decay.ms}{stress decay per ms}

\item{decay.perc.s}{stress decay percentage per second

Note that 1.0 means 100% decay per second}

\item{stress.init}{stress to start off with}

\item{bin.size}{the bins size for model smoothing

The smaller the bin size the more suspectible the model will be to sudden
spikes.}

\item{sim.disable}{disables stress simulation. Improves performance}
}
\description{
Simulates the Models through types
}
