% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSim.R
\name{model.sim}
\alias{model.sim}
\title{Simulates the Models through types}
\usage{
model.sim(m.jck, m.mtn, m.dns, m.mnp, jck.pow, mtn.pow, dns.pow,
  decay.ms = 0.5, stress.init = 0, bin.size = 5000)
}
\arguments{
\item{m.jck}{model generated by model.jackInv}

\item{m.mtn}{model generated by model.motion}

\item{m.dns}{model generated by model.density}

\item{jck.pow}{power factor for jck}

\item{mtn.pow}{power factor for mtn}

\item{dns.pow}{power factor for dns}

\item{decay.ms}{stress decay per ms}

\item{stress.init}{stress to start off with}

\item{bin.size}{the bins size for model smoothing
The smaller the bin size the more suspectible the
model will be to sudden spikes.}
}
\description{
Simulates the Models through types
}
