% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_sim.R
\name{stressSim}
\alias{stressSim}
\title{Simulates the Stress through types}
\usage{
stressSim(chart, stress.init = 0, decay_ms = 0.01)
}
\arguments{
\item{chart}{Chart generated by `f.stress.mapper`}

\item{stress.init}{The **numeric** of 'stress' to
initialize the simulation with}
}
\description{

}
